fix: figure out if we need to install call the update pypi prefix before we use the interpreter
