+++ src/install_pypi.rs
@@ -1,4 +1,3 @@
-use crate::environment::PythonStatus;
 use crate::prefix::Prefix;
 use crate::project::manifest::pypi_options::PypiOptions;
 use crate::uv_reporter::{UvReporter, UvReporterOptions};
@@ -534,54 +533,6 @@ fn whats_the_plan<'a>(
     })
 }
 
-/// If the python interpreter is outdated, we need to uninstall all outdated site packages.
-/// from the old interpreter.
-async fn uninstall_outdated_site_packages(site_packages: &Path) -> miette::Result<()> {
-    // Check if the old interpreter is outdated
-    let mut installed = vec![];
-    for entry in std::fs::read_dir(site_packages).into_diagnostic()? {
-        let entry = entry.into_diagnostic()?;
-        if entry.file_type().into_diagnostic()?.is_dir() {
-            let path = entry.path();
-
-            let installed_dist = InstalledDist::try_from_path(&path);
-            let Ok(installed_dist) = installed_dist else {
-                continue;
-            };
-
-            if let Some(installed_dist) = installed_dist {
-                // If we can't get the installer, we can't be certain that we have installed it
-                let installer = match installed_dist.installer() {
-                    Ok(installer) => installer,
-                    Err(e) => {
-                        tracing::warn!(
-                            "could not get installer for {}: {}, will not remove distribution",
-                            installed_dist.name(),
-                            e
-                        );
-                        continue;
-                    }
-                };
-
-                // Only remove if have actually installed it
-                // by checking the installer
-                if installer.unwrap_or_default() == PIXI_UV_INSTALLER {
-                    installed.push(installed_dist);
-                }
-            }
-        }
-    }
-
-    // Uninstall all packages in old site-packages directory
-    for dist_info in installed {
-        let _summary = uv_installer::uninstall(&dist_info)
-            .await
-            .expect("uninstallation of old site-packages failed");
-    }
-
-    Ok(())
-}
-
 /// Result of resolving editables
 /// we need to store the temp_dir until the install is finished
 struct EditablesWithTemp {
@@ -731,7 +682,7 @@ pub async fn update_python_distributions(
     prefix: &Prefix,
     conda_package: &[RepoDataRecord],
     python_packages: &[CombinedPypiPackageData],
-    status: &PythonStatus,
+    python_interpreter_path: &Path,
     system_requirements: &SystemRequirements,
     uv_context: UvResolutionContext,
     pypi_options: &PypiOptions,
@@ -739,18 +690,6 @@ pub async fn update_python_distributions(
     platform: Platform,
 ) -> miette::Result<()> {
     let start = std::time::Instant::now();
-    let Some(python_info) = status.current_info() else {
-        // No python interpreter in the environment, so there is nothing to do here.
-        return Ok(());
-    };
-
-    // If we have changed interpreter, we need to uninstall all site-packages from the old interpreter
-    if let PythonStatus::Changed { old, new: _ } = status {
-        let site_packages_path = prefix.root().join(&old.site_packages_path);
-        if site_packages_path.exists() {
-            uninstall_outdated_site_packages(&site_packages_path).await?;
-        }
-    }
 
     // Determine the current environment markers.
     let python_record = conda_package
@@ -788,7 +727,7 @@ pub async fn update_python_distributions(
     let in_memory_index = InMemoryIndex::default();
     let config_settings = ConfigSettings::default();
 
-    let python_location = prefix.root().join(&python_info.path);
+    let python_location = prefix.root().join(python_interpreter_path);
     let interpreter = Interpreter::query(&python_location, &uv_context.cache).into_diagnostic()?;
 
     tracing::debug!("[Install] Using Python Interpreter: {:?}", interpreter);
